Update app.js
After further examination, it breaks master to have setUnits, setColumns and setDests as the functions defined in the properties of the <Home.../> in app.js.  Unless they are called in Home.jsx with setColumns, setUnits, etc. (which they are not now) then the selections will not work, as they're using getUnits, getColumns.  It would involve changing the method calls in many places/lines of code in Home.jsx, so that the name of these defined methods are setSomething instead of getSomething.  While these methods are both setting and getting, they are doing more in the way of getting than setting.

This changes everything back to getSomething in app.js so this.props.Method can be used in the way it is written in Home.jsx.